separate a function for starting the rqlite server
