+++ supernode/.gitignore
@@ -1,4 +1,4 @@
 *.log
 supernode
-data
 scripts
+data
