+++ supernode/examples/configs/testnet-4446.yml
@@ -9,3 +9,10 @@ node:
     # `listen_address` and `port` must match to `extAddress` from masternode.conf
     listen_addresses: "127.0.0.1"
     port: 4446
+  rqlite:
+    http_address: "localhost:4005"
+    raft_address: "localhost:4006"
+    node_id: "3"
+    data_dir: "data/node3"
+    raft_open_timeout: "5s"
+    join_address: "http://localhost:4001"
